Moved integration/containers to pkg/testing/containers
