+++ integration/ig/non-k8s/trace_network_test.go
@@ -19,8 +19,8 @@ import (
 	"testing"
 
 	. "github.com/inspektor-gadget/inspektor-gadget/integration"
-	"github.com/inspektor-gadget/inspektor-gadget/integration/containers"
 	networkTypes "github.com/inspektor-gadget/inspektor-gadget/pkg/gadgets/trace/network/types"
+	"github.com/inspektor-gadget/inspektor-gadget/pkg/testing/containers"
 	eventtypes "github.com/inspektor-gadget/inspektor-gadget/pkg/types"
 )
